refactor: reorganize test utilities and update import paths
