+++ test/e2e/e2e_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/mycel-labs/transferable-account/src/go/framework"
 	pb "github.com/mycel-labs/transferable-account/src/go/pb/api/v1"
-	"github.com/mycel-labs/transferable-account/test/testutil"
+	testutil "github.com/mycel-labs/transferable-account/test/utils"
 )
 
 var (
